+++ insights/get_filters_from_insights.py
@@ -1,7 +1,7 @@
-from handle_user_prompt import get_prompt_insights
-from handle_user_prompt import get_prompt
-from userPreference import get_user_purchase_insight
-from uniqueValues import brand_name_array, color_array, article_type_array,occasion_array
+from prompt_insights import get_prompt_insights
+from prompt_insights import get_prompt
+from insights.user_purchase_insights import get_user_purchase_insight
+from utils.uniqueValues import brand_name_array, color_array, article_type_array,occasion_array
 
 
 def categorize_filters(insights, unique_array_dict):
@@ -40,22 +40,21 @@ def analyse_user_bio_data():
     gender = "women"
     return gender
 
-def analyse_user_purchase_insights():
-    user_purchase_csv = 'dataset/user_history_data/gwen.csv'
+def analyse_user_purchase_insights(user_purchase_csv):
     user_purchase_insights = get_user_purchase_insight(user_purchase_csv)
     # top_wear, bottom_wear, foot_wear, accessories = user_purchase_insights
     print("user_purchase_insights:", user_purchase_insights)
     hard_filters_purchase, soft_filters_purchase = categorize_filters(user_purchase_insights, unique_array_dict)
     print("soft_filters_purchase:", soft_filters_purchase)
     print("hard_filters_purchase:", hard_filters_purchase)
-
+    return hard_filters_purchase, soft_filters_purchase
 # For user's purchase insights
 
 
 
 
-def analyse_user_prompt_insights():
-    user_prompt = get_prompt()
+def analyse_user_prompt_insights(user_prompt):
+
     top_wear, bottom_wear, foot_wear, accessories = get_prompt_insights(user_prompt=user_prompt)
     user_prompt_insights = [top_wear, bottom_wear, foot_wear, accessories]
     print("user_prompt_insights:", user_prompt_insights)
@@ -63,14 +62,15 @@ def analyse_user_prompt_insights():
     hard_filters_prompt, soft_filters_prompt = categorize_filters(user_prompt_insights, unique_array_dict)
     print("soft_filters_prompt:", soft_filters_prompt)
     print("hard_filters_prompt:", hard_filters_prompt)
-
-
+    return hard_filters_prompt, soft_filters_prompt
 
 
 if __name__ == "__main__":
     print("\n\nuser_purchase_insights:::: ")
-    analyse_user_purchase_insights()
+    user_purchase_csv = 'dataset/user_history_data/gwen.csv'
+    analyse_user_purchase_insights(user_purchase_csv)
     print("\n\n***************************\n")
     print("\n\nuser_prompt_insights:::: ")
-    analyse_user_prompt_insights()
+    user_prompt = get_prompt()
+    analyse_user_prompt_insights(user_prompt)
     print("\n\n***************************")
