+++ insights/prompt_insights.py
